@@ -73,7 +73,7 @@ allprojects {
     }
   }
 
-  sourceCompatibility = 1.8
+  sourceCompatibility = 11
   compileJava {
     options.compilerArgs << '-parameters' << '-Xlint:deprecation'
   }
