@@ -5,7 +5,7 @@ plugins {
 
 apply plugin: 'net.ltgt.apt'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 11
 
 repositories {
   mavenCentral()
@@ -17,14 +17,14 @@ dependencies {
   compile group: 'cz.jirutka.validator', name: 'validator-collection', version: '2.2.0'
 
   //   The below fixes https://nvd.nist.gov/vuln/detail/CVE-2019-12384 while waiting for spring to pull new version
-  compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.10.2'
+  compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.10.3'
 
   annotationProcessor group: 'org.projectlombok', name: 'lombok', version: '1.18.12'
   compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.10'
 
   testCompile project(':domain-sample-data')
   testCompile group: 'junit', name: 'junit', version: '4.13'
-  testCompile group: 'org.mockito', name: 'mockito-core', version: '3.2.4'
+  testCompile group: 'org.mockito', name: 'mockito-core', version: '3.3.3'
   testCompile group: 'org.assertj', name: 'assertj-core', version: '3.15.0'
   testCompile group: 'javax.el', name: 'javax.el-api', version: '3.0.0'
   testCompile group: 'org.glassfish.web', name: 'javax.el', version: '2.2.6'
