@@ -1,6 +1,5 @@
 package uk.gov.hmcts.cmc.ccd.util;
 
-import org.hamcrest.core.Is;
 import org.junit.Test;
 import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
 import uk.gov.hmcts.cmc.ccd.sample.data.SampleCCDDefendant;
@@ -18,11 +17,7 @@ import java.time.LocalDateTime;
 import java.util.Arrays;
 
 import static java.util.Collections.singletonList;
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.cmc.ccd.sample.data.SampleCCDClaimSubmissionOperationIndicators.CCDClaimSubmissionOperationIndicatorsWithPinSuccess;
 import static uk.gov.hmcts.cmc.ccd.util.MapperUtil.getMediationOutcome;
 import static uk.gov.hmcts.cmc.ccd.util.MapperUtil.hasPaperResponse;
@@ -34,7 +29,7 @@ public class MapperUtilTest {
     public void caseNameNotNull() {
         Claim sampleClaim = SampleClaim.getDefault();
         String caseName = toCaseName.apply(sampleClaim);
-        assertNotNull(caseName);
+        assertThat(caseName).isNotNull();
     }
 
     @Test
@@ -53,8 +48,7 @@ public class MapperUtilTest {
         ).build();
 
         String caseName = toCaseName.apply(claimWithMultiClaimant);
-        assertNotNull(caseName);
-        assertThat(caseName, is("Brexiter + others Vs Mrs. Theresa May"));
+        assertThat(caseName).isEqualTo("Brexiter + others Vs Mrs. Theresa May");
 
     }
 
@@ -79,8 +73,7 @@ public class MapperUtilTest {
         ).build();
 
         String caseName = toCaseName.apply(claimWithMultiDefendant);
-        assertNotNull(caseName);
-        assertThat(caseName, is("Brexiter Vs Mrs. Theresa May + others"));
+        assertThat(caseName).isEqualTo("Brexiter Vs Mrs. Theresa May + others");
 
     }
 
@@ -90,9 +83,9 @@ public class MapperUtilTest {
         Claim claimWithClaimantSoleTrader = Claim.builder().claimData(
             SampleClaimData.builder(
                 singletonList(SampleParty.builder()
-                        .withName("Georgina Hammersmith")
-                        .withBusinessName("EuroStar")
-                        .soleTrader()),
+                    .withName("Georgina Hammersmith")
+                    .withBusinessName("EuroStar")
+                    .soleTrader()),
                 singletonList(SampleTheirDetails.builder()
                     .withTitle("Mr.")
                     .withFirstName("Boris")
@@ -102,32 +95,30 @@ public class MapperUtilTest {
         ).build();
 
         String caseName = toCaseName.apply(claimWithClaimantSoleTrader);
-        assertNotNull(caseName);
-        assertThat(caseName, is("Georgina Hammersmith T/A EuroStar Vs Mr. Boris Johnson"));
+        assertThat(caseName).isEqualTo("Georgina Hammersmith T/A EuroStar Vs Mr. Boris Johnson");
 
     }
 
     @Test
     public void caseNameWhenDefendantIsSoleTrader() {
         Claim claimWithDefendantSoleTrader = Claim.builder().claimData(
-                SampleClaimData.builder(
-                        singletonList(SampleParty.builder()
-                                .withTitle("Mrs.")
-                                .withName("Boi May")
-                                .individual()),
-                        singletonList(SampleTheirDetails.builder()
-                                .withTitle("Mr.")
-                                .withFirstName("Boris")
-                                .withLastName("Johnson")
-                                .withBusinessName("Uberflip")
-                                .soleTraderDetails()
-                        )
-                ).build()
+            SampleClaimData.builder(
+                singletonList(SampleParty.builder()
+                    .withTitle("Mrs.")
+                    .withName("Boi May")
+                    .individual()),
+                singletonList(SampleTheirDetails.builder()
+                    .withTitle("Mr.")
+                    .withFirstName("Boris")
+                    .withLastName("Johnson")
+                    .withBusinessName("Uberflip")
+                    .soleTraderDetails()
+                )
+            ).build()
         ).build();
 
         String caseName = toCaseName.apply(claimWithDefendantSoleTrader);
-        assertNotNull(caseName);
-        assertThat(caseName, is("Boi May Vs Mr. Boris Johnson T/A Uberflip"));
+        assertThat(caseName).isEqualTo("Boi May Vs Mr. Boris Johnson T/A Uberflip");
 
     }
 
@@ -135,23 +126,22 @@ public class MapperUtilTest {
     public void caseNameWhenBothAreSoleTrader() {
 
         Claim claimWithBothAsSoleTrader = Claim.builder().claimData(
-                SampleClaimData.builder(
-                        singletonList(SampleParty.builder()
-                                .withName("Georgina Hammersmith")
-                                .withBusinessName("EuroStar")
-                                .soleTrader()),
-                        singletonList(SampleTheirDetails.builder()
-                                .withTitle("Mr.")
-                                .withFirstName("Boris")
-                                .withLastName("Johnson")
-                                .withBusinessName("Haberdashery")
-                                .soleTraderDetails())
-                ).build()
+            SampleClaimData.builder(
+                singletonList(SampleParty.builder()
+                    .withName("Georgina Hammersmith")
+                    .withBusinessName("EuroStar")
+                    .soleTrader()),
+                singletonList(SampleTheirDetails.builder()
+                    .withTitle("Mr.")
+                    .withFirstName("Boris")
+                    .withLastName("Johnson")
+                    .withBusinessName("Haberdashery")
+                    .soleTraderDetails())
+            ).build()
         ).build();
 
         String caseName = toCaseName.apply(claimWithBothAsSoleTrader);
-        assertNotNull(caseName);
-        assertThat(caseName, is("Georgina Hammersmith T/A EuroStar Vs Mr. Boris Johnson T/A Haberdashery"));
+        assertThat(caseName).isEqualTo("Georgina Hammersmith T/A EuroStar Vs Mr. Boris Johnson T/A Haberdashery");
 
     }
 
@@ -171,8 +161,7 @@ public class MapperUtilTest {
             .build();
 
         String caseName = MapperUtil.toCaseName.apply(claimWithResponse);
-        assertNotNull(caseName);
-        assertThat(caseName, is("Versace T/A Versace Vs French Connection UK"));
+        assertThat(caseName).isEqualTo("Versace T/A Versace Vs French Connection UK");
     }
 
     @Test
@@ -183,7 +172,7 @@ public class MapperUtilTest {
             SampleData.getCCDCitizenCaseWithRespondent(
                 SampleCCDDefendant.withMediationAgreementDate(mediationSettledTime).build());
 
-        assertThat(getMediationOutcome(ccdCase), Is.is(MediationOutcome.SUCCEEDED));
+        assertThat(getMediationOutcome(ccdCase)).isEqualTo(MediationOutcome.SUCCEEDED);
     }
 
     @Test
@@ -191,7 +180,7 @@ public class MapperUtilTest {
         CCDCase ccdCase =
             SampleData.getCCDCitizenCaseWithRespondent(SampleCCDDefendant.withMediationFailureReason().build());
 
-        assertThat(getMediationOutcome(ccdCase), Is.is(MediationOutcome.FAILED));
+        assertThat(getMediationOutcome(ccdCase)).isEqualTo(MediationOutcome.FAILED);
     }
 
     @Test
@@ -199,7 +188,7 @@ public class MapperUtilTest {
         CCDCase ccdCase =
             SampleData.getCCDCitizenCaseWithOperationIndicators(CCDClaimSubmissionOperationIndicatorsWithPinSuccess);
 
-        assertNull(getMediationOutcome(ccdCase));
+        assertThat(getMediationOutcome(ccdCase)).isNull();
     }
 
     @Test
@@ -208,7 +197,7 @@ public class MapperUtilTest {
             SampleData.getCCDCitizenCaseWithOperationIndicators(CCDClaimSubmissionOperationIndicatorsWithPinSuccess);
 
         YesNoOption result = hasPaperResponse.apply(ccdCase);
-        assertEquals(YesNoOption.NO, result);
+        assertThat(result).isEqualTo(YesNoOption.NO);
 
     }
 
@@ -218,7 +207,7 @@ public class MapperUtilTest {
             SampleData.withPaperResponseFromStaffUploadedDoc();
 
         YesNoOption result = hasPaperResponse.apply(ccdCase);
-        assertEquals(YesNoOption.YES, result);
+        assertThat(result).isEqualTo(YesNoOption.YES);
 
     }
 
@@ -228,7 +217,7 @@ public class MapperUtilTest {
             SampleData.withPaperResponseFromScannedDoc();
 
         YesNoOption result = hasPaperResponse.apply(ccdCase);
-        assertEquals(YesNoOption.YES, result);
+        assertThat(result).isEqualTo(YesNoOption.YES);
 
     }
 
