@@ -17,7 +17,8 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 @EnableFeignClients(basePackages =
     {"uk.gov.hmcts.cmc.claimstore",
         "uk.gov.hmcts.reform.authorisation",
-        "uk.gov.hmcts.reform.sendletter"
+        "uk.gov.hmcts.reform.sendletter",
+        "uk.gov.hmcts.reform.ccd.client"
     })
 public class ClaimStoreApplication {
 
