@@ -1,5 +1,6 @@
 package uk.gov.hmcts.cmc.claimstore.repositories;
 
+import org.elasticsearch.index.query.QueryBuilder;
 import uk.gov.hmcts.cmc.claimstore.idam.models.User;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 
@@ -7,9 +8,10 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface CaseSearchApi {
+
     List<Claim> getMediationClaims(String authorisation, LocalDate mediationAgreementDate);
 
-    List<Claim> getClaimsPastIntentionToProceed(User user, LocalDate responseDate);
+    List<Claim> getClaims(User user, QueryBuilder queryBuilder);
 
     List<Claim> getClaimsWithDefaultCCJ(User user, LocalDate ccjRequestedDate);
 }
