@@ -24,7 +24,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.DRAW_ORDER;
 import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.JUDGE;
 import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.LEGAL_ADVISOR;
 
@@ -32,7 +31,7 @@ import static uk.gov.hmcts.cmc.claimstore.services.ccd.Role.LEGAL_ADVISOR;
 @ConditionalOnProperty(prefix = "doc_assembly", name = "url")
 public class DrawOrderCallbackHandler extends CallbackHandler {
     private static final List<Role> ROLES = ImmutableList.of(LEGAL_ADVISOR, JUDGE);
-    private static final List<CaseEvent> EVENTS = Collections.singletonList(DRAW_ORDER);
+    private static final List<CaseEvent> EVENTS = Collections.singletonList(CaseEvent.DRAW_ORDER);
     private static final String DRAFT_ORDER_DOC = "draftOrderDoc";
 
     private final CaseDetailsConverter caseDetailsConverter;
