@@ -8,6 +8,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.cmc.ccd.domain.CCDCase;
+import uk.gov.hmcts.cmc.ccd.domain.CCDCollectionElement;
 import uk.gov.hmcts.cmc.ccd.domain.CCDDocument;
 import uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption;
 import uk.gov.hmcts.cmc.ccd.domain.CaseEvent;
@@ -39,6 +40,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.NO;
 import static uk.gov.hmcts.cmc.ccd.domain.CCDYesNoOption.YES;
@@ -67,7 +69,11 @@ public class OrderCreator {
     private static final String DYNAMIC_LIST_CODE = "code";
     private static final String DYNAMIC_LIST_LABEL = "label";
     private static final String DYNAMIC_LIST_ITEMS = "list_items";
+    private static final String DYNAMIC_LIST_SELECTED_VALUE = "value";
     private static final String GRANT_EXPERT_REPORT_PERMISSION = "grantExpertReportPermission";
+    private static final String EXPERT_REPORT_INSTRUCTION = "expertReportInstruction";
+    private static final String OTHER_DIRECTIONS = "otherDirections";
+    private static final String ESTIMATED_HEARING_DURATION = "estimatedHearingDuration";
 
     private final LegalOrderGenerationDeadlinesCalculator legalOrderGenerationDeadlinesCalculator;
     private final CaseDetailsConverter caseDetailsConverter;
@@ -96,32 +102,73 @@ public class OrderCreator {
         logger.info("Order creator: pre populating order fields");
         CallbackRequest callbackRequest = callbackParams.getRequest();
 
-        Map<String, Object> data = new HashMap<>();
-        data.put(DIRECTION_LIST, ImmutableList.of(DOCUMENTS.name(), EYEWITNESS.name()));
-
         Claim claim = caseDetailsConverter.extractClaim(callbackRequest.getCaseDetails());
         CCDCase ccdCase = caseDetailsConverter.extractCCDCase(callbackRequest.getCaseDetails());
+
+        Map<String, Object> data = new HashMap<>();
+        data.put(DIRECTION_LIST, chooseItem(ccdCase.getDirectionList(), ImmutableList.of(DOCUMENTS, EYEWITNESS)));
+
         addCourtData(claim, ccdCase, data);
 
         LocalDate deadline = legalOrderGenerationDeadlinesCalculator.calculateOrderGenerationDeadlines();
-        data.put(DOC_UPLOAD_DEADLINE, deadline);
-        data.put(EYEWITNESS_UPLOAD_DEADLINE, deadline);
-        data.put(DOC_UPLOAD_FOR_PARTY, BOTH.name());
-        data.put(EYEWITNESS_UPLOAD_FOR_PARTY, BOTH.name());
-        data.put(PAPER_DETERMINATION, NO.name());
+        data.put(DOC_UPLOAD_DEADLINE, chooseItem(ccdCase.getDocUploadDeadline(), deadline));
+        data.put(EYEWITNESS_UPLOAD_DEADLINE, chooseItem(ccdCase.getEyewitnessUploadDeadline(), deadline));
+        data.put(DOC_UPLOAD_FOR_PARTY, chooseItem(ccdCase.getDocUploadForParty(), BOTH));
+        data.put(EYEWITNESS_UPLOAD_FOR_PARTY, chooseItem(ccdCase.getEyewitnessUploadForParty(), BOTH));
+        data.put(PAPER_DETERMINATION, chooseItem(ccdCase.getPaperDetermination(), NO));
+        data.put(ESTIMATED_HEARING_DURATION, ccdCase.getEstimatedHearingDuration());
 
-        if (hasExpertsAtCaseLevel(callbackParams)) {
-            data.put(HEARING_COURT, buildCourtsList(getPilot(callbackParams), claim.getCreatedAt()));
-        }
+        data.put(OTHER_DIRECTIONS, ccdCase.getOtherDirections());
 
         if (hasDynamicCourts(callbackParams)) {
-            data.put(GRANT_EXPERT_REPORT_PERMISSION, NO);
+            data.put(GRANT_EXPERT_REPORT_PERMISSION, chooseItem(ccdCase.getGrantExpertReportPermission(), NO));
+            data.put(EXPERT_REPORT_INSTRUCTION, chooseItem(ccdCase.getExpertReportInstruction(), null));
+        }
+
+        if (hasExpertsAtCaseLevel(callbackParams)) {
+            data.put(HEARING_COURT,
+                buildCourtsList(getPilot(callbackParams), claim.getCreatedAt(), ccdCase.getHearingCourtName()));
+
+            if (ccdCase.getGrantExpertReportPermission() == null
+                && (ccdCase.getExpertReportPermissionPartyGivenToClaimant() != null
+                || ccdCase.getExpertReportPermissionPartyGivenToDefendant() != null)) {
+
+                data.put(GRANT_EXPERT_REPORT_PERMISSION, hasExpertPermission(ccdCase) ? YES : NO);
+
+                String instructions = Stream.of(extractInstructions(ccdCase.getExpertReportInstructionClaimant()),
+                    extractInstructions(ccdCase.getExpertReportInstructionDefendant()))
+                    .filter(StringUtils::isNotBlank)
+                    .collect(Collectors.joining(", "));
+
+                data.put(EXPERT_REPORT_INSTRUCTION, instructions);
+
+            }
         }
 
         return AboutToStartOrSubmitCallbackResponse
             .builder()
             .data(data)
             .build();
+    }
+
+    private static <T> T chooseItem(T value, T defaultValue) {
+        return value != null ? value : defaultValue;
+    }
+
+    private static <T> List<T> chooseItem(List<T> list, List<T> defaultList) {
+        return list != null && !list.isEmpty() ? list : defaultList;
+    }
+
+    private boolean hasExpertPermission(CCDCase ccdCase) {
+        return ccdCase.getExpertReportPermissionPartyGivenToClaimant().toBoolean()
+            || ccdCase.getExpertReportPermissionPartyGivenToDefendant().toBoolean();
+    }
+
+    private String extractInstructions(List<CCDCollectionElement<String>> expertReportInstructions) {
+        return expertReportInstructions.stream()
+            .map(CCDCollectionElement::getValue)
+            .filter(StringUtils::isNotBlank)
+            .collect(Collectors.joining(", "));
     }
 
     public CallbackResponse generateOrder(CallbackParams callbackParams) {
@@ -207,7 +254,7 @@ public class OrderCreator {
             && StringUtils.isNotBlank(directionsQuestionnaire.getExpertEvidenceToExamine());
     }
 
-    private Map<String, List<Map<String, String>>> buildCourtsList(Pilot pilot, LocalDateTime claimCreatedDate) {
+    private Map<String, Object> buildCourtsList(Pilot pilot, LocalDateTime claimCreatedDate, String hearingCourtName) {
         List<Map<String, String>> listItems = pilotCourtService.getPilotHearingCourts(pilot, claimCreatedDate).stream()
             .sorted(Comparator.comparing(HearingCourt::getName))
             .map(hearingCourt -> {
@@ -216,12 +263,30 @@ public class OrderCreator {
             })
             .collect(Collectors.toList());
 
+        ImmutableMap<String, String> otherCourtItem = ImmutableMap.of(DYNAMIC_LIST_CODE,
+            PilotCourtService.OTHER_COURT_ID, DYNAMIC_LIST_LABEL, "Other Court");
+
         if (pilot == Pilot.JDDO) {
-            listItems.add(ImmutableMap.of(DYNAMIC_LIST_CODE, PilotCourtService.OTHER_COURT_ID,
-                DYNAMIC_LIST_LABEL, "Other Court"));
+            listItems.add(otherCourtItem);
         }
 
-        return ImmutableMap.of(DYNAMIC_LIST_ITEMS, listItems);
+        Map<String, Object> hearingCourtListDefinition = new HashMap<>();
+        hearingCourtListDefinition.put(DYNAMIC_LIST_ITEMS, listItems);
+
+        if (StringUtils.isBlank(hearingCourtName)) {
+            return hearingCourtListDefinition;
+        }
+
+        Optional<Map<String, String>> selectedCourt =
+            listItems.stream().filter(s -> s.get(DYNAMIC_LIST_LABEL).equals(hearingCourtName)).findFirst();
+
+        if (selectedCourt.isPresent()) {
+            hearingCourtListDefinition.put(DYNAMIC_LIST_SELECTED_VALUE, selectedCourt.get());
+        } else if (pilot == Pilot.JDDO) {
+            hearingCourtListDefinition.put(DYNAMIC_LIST_SELECTED_VALUE, otherCourtItem);
+        }
+
+        return hearingCourtListDefinition;
     }
 
     private Pilot getPilot(CallbackParams callbackParams) {
