@@ -39,7 +39,7 @@ public interface SampleQueryConstants {
         + "  }\n"
         + "}}";
 
-    String stayableCaseQuery = "{\"size\": 1000,\"query\": {\n"
+    String stayableCaseQuery = "{\n"
         + "  \"bool\" : {\n"
         + "    \"must\" : [\n"
         + "      {\n"
@@ -64,8 +64,50 @@ public interface SampleQueryConstants {
         + "      {\n"
         + "        \"range\" : {\n"
         + "          \"data.submittedOn\" : {\n"
-        + "            \"from\" : \"2019-09-09T03:12:00.000Z\",\n"
+        + "            \"from\" : \"2019-09-09T15:12:00.000Z\",\n"
         + "            \"to\" : null,\n"
+        + "            \"include_lower\" : true,\n"
+        + "            \"include_upper\" : true,\n"
+        + "            \"boost\" : 1.0\n"
+        + "          }\n"
+        + "        }\n"
+        + "      }\n"
+        + "    ],\n"
+        + "    \"must_not\" : [\n"
+        + "      {\n"
+        + "        \"exists\" : {\n"
+        + "          \"field\" : \"data.respondents.value.paidInFullDate\",\n"
+        + "          \"boost\" : 1.0\n"
+        + "        }\n"
+        + "      },\n"
+        + "      {\n"
+        + "        \"exists\" : {\n"
+        + "          \"field\" : \"data.respondents.value.claimantResponse.submittedOn\",\n"
+        + "          \"boost\" : 1.0\n"
+        + "        }\n"
+        + "      }\n"
+        + "    ],\n"
+        + "    \"adjust_pure_negative\" : true,\n"
+        + "    \"boost\" : 1.0\n"
+        + "  }\n"
+        + "}";
+
+    String waitingTransferQuery = "{\n"
+        + "  \"bool\" : {\n"
+        + "    \"must\" : [\n"
+        + "      {\n"
+        + "        \"term\" : {\n"
+        + "          \"state\" : {\n"
+        + "            \"value\" : \"orderdrawn\",\n"
+        + "            \"boost\" : 1.0\n"
+        + "          }\n"
+        + "        }\n"
+        + "      },\n"
+        + "      {\n"
+        + "        \"range\" : {\n"
+        + "          \"data.directionOrder.createdOn\" : {\n"
+        + "            \"from\" : null,\n"
+        + "            \"to\" : \"2019-07-07\",\n"
         + "            \"include_lower\" : true,\n"
         + "            \"include_upper\" : true,\n"
         + "            \"boost\" : 1.0\n"
@@ -76,7 +118,7 @@ public interface SampleQueryConstants {
         + "    \"adjust_pure_negative\" : true,\n"
         + "    \"boost\" : 1.0\n"
         + "  }\n"
-        + "}}";
+        + "}";
 
     String defaultCCJCases10DaysBefore = "{\"size\": 1000,\"query\": {\n"
         + "  \"bool\" : {\n"
