@@ -19,13 +19,19 @@ import uk.gov.hmcts.cmc.domain.models.claimantresponse.ResponseAcceptation;
 import uk.gov.hmcts.cmc.domain.models.claimantresponse.ResponseRejection;
 import uk.gov.hmcts.cmc.domain.models.directionsquestionnaire.DirectionsQuestionnaire;
 import uk.gov.hmcts.cmc.domain.models.response.DefenceType;
+import uk.gov.hmcts.cmc.domain.models.response.FullDefenceResponse;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaimantResponse;
+import uk.gov.hmcts.cmc.domain.models.sampledata.SampleDirectionsQuestionnaire;
+import uk.gov.hmcts.cmc.domain.models.sampledata.SampleHearingLocation;
 import uk.gov.hmcts.cmc.domain.models.sampledata.SampleResponse;
 
 import java.math.BigDecimal;
+import java.time.Clock;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
 import java.util.Optional;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -33,9 +39,9 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.ASSIGNING_FOR_JUDGE_DIRECTIONS;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.ASSIGNING_FOR_LEGAL_ADVISOR_DIRECTIONS;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.LIFT_STAY;
 import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.REFERRED_TO_MEDIATION;
@@ -43,16 +49,19 @@ import static uk.gov.hmcts.cmc.ccd.domain.CaseEvent.STAY_CLAIM;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsights.REFERENCE_NUMBER;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.BOTH_OPTED_IN_FOR_MEDIATION_PILOT;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.BOTH_OPTED_IN_FOR_NON_MEDIATION_PILOT;
+import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.BOTH_PARTIES_OFFLINE_DQ;
+import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.BOTH_PARTIES_ONLINE_DQ;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.CLAIMANT_OPTED_OUT_FOR_MEDIATION_PILOT;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.CLAIMANT_OPTED_OUT_FOR_NON_MEDIATION_PILOT;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.CLAIMANT_RESPONSE_ACCEPTED;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.CLAIM_STAYED;
+import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.JDDO_PILOT_ELIGIBLE;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.LA_PILOT_ELIGIBLE;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.MEDIATION_NON_PILOT_ELIGIBLE;
 import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.MEDIATION_PILOT_ELIGIBLE;
-import static uk.gov.hmcts.cmc.claimstore.appinsights.AppInsightsEvent.NON_LA_CASES;
 import static uk.gov.hmcts.cmc.claimstore.utils.VerificationModeUtils.once;
 import static uk.gov.hmcts.cmc.domain.models.ClaimFeatures.DQ_FLAG;
+import static uk.gov.hmcts.cmc.domain.models.ClaimFeatures.JUDGE_PILOT_FLAG;
 import static uk.gov.hmcts.cmc.domain.models.ClaimFeatures.LA_PILOT_FLAG;
 import static uk.gov.hmcts.cmc.domain.models.ClaimFeatures.MEDIATION_PILOT;
 import static uk.gov.hmcts.cmc.domain.models.response.YesNoOption.YES;
@@ -62,6 +71,8 @@ import static uk.gov.hmcts.cmc.domain.models.sampledata.SampleClaim.EXTERNAL_ID;
 public class ClaimantResponseServiceTest {
 
     private static final String AUTHORISATION = "Bearer: aaa";
+    private static final ZonedDateTime TODAY_ZONED = LocalDate.of(2019, 1, 2).atStartOfDay(ZoneOffset.UTC);
+    private static final LocalDateTime TODAY = TODAY_ZONED.toLocalDateTime();
 
     private ClaimantResponseService claimantResponseService;
 
@@ -86,6 +97,9 @@ public class ClaimantResponseServiceTest {
     @Mock
     private DirectionsQuestionnaireDeadlineCalculator directionsQuestionnaireDeadlineCalculator;
 
+    @Mock
+    private Clock clock;
+
     @Before
     public void setUp() {
         claimantResponseService = new ClaimantResponseService(
@@ -96,8 +110,11 @@ public class ClaimantResponseServiceTest {
             eventProducer,
             formaliseResponseAcceptanceService,
             directionsQuestionnaireService,
-            directionsQuestionnaireDeadlineCalculator
+            directionsQuestionnaireDeadlineCalculator,
+            clock
         );
+        when(clock.instant()).thenReturn(TODAY_ZONED.toInstant());
+        when(clock.getZone()).thenReturn(ZoneOffset.UTC);
     }
 
     @Test
@@ -121,10 +138,11 @@ public class ClaimantResponseServiceTest {
 
         inOrder.verify(caseRepository, once()).saveClaimantResponse(any(Claim.class), eq(claimantResponse), any());
         inOrder.verify(eventProducer, once()).createClaimantResponseEvent(any(Claim.class), anyString());
-        inOrder.verify(appInsights, once()).trackEvent(eq(NON_LA_CASES),
-            eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
-        verify(formaliseResponseAcceptanceService, times(0))
-            .formalise(any(), any(), anyString());
+        verify(appInsights, once()).trackEvent(eq(BOTH_PARTIES_OFFLINE_DQ),
+                eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
+
+        verify(formaliseResponseAcceptanceService, never()).formalise(any(), any(), anyString());
+
         verify(caseRepository, never()).saveCaseEvent(AUTHORISATION, claim, ASSIGNING_FOR_LEGAL_ADVISOR_DIRECTIONS);
         verify(caseRepository, never()).saveCaseEvent(AUTHORISATION, claim, REFERRED_TO_MEDIATION);
     }
@@ -300,17 +318,16 @@ public class ClaimantResponseServiceTest {
         when(claimService.getClaimByExternalId(eq(EXTERNAL_ID), eq(AUTHORISATION))).thenReturn(claim);
         when(caseRepository.saveClaimantResponse(any(Claim.class), any(ResponseRejection.class), eq(AUTHORISATION)))
             .thenReturn(claim);
-        when(directionsQuestionnaireDeadlineCalculator
-            .calculateDirectionsQuestionnaireDeadline(any(LocalDateTime.class)))
+        when(directionsQuestionnaireDeadlineCalculator.calculateDirectionsQuestionnaireDeadline(eq(TODAY)))
             .thenReturn(dqDeadline);
 
         claimantResponseService.save(EXTERNAL_ID, claim.getSubmitterId(), claimantResponse, AUTHORISATION);
 
         verify(caseRepository).saveClaimantResponse(any(Claim.class), eq(claimantResponse), any());
-        verify(directionsQuestionnaireDeadlineCalculator)
-            .calculateDirectionsQuestionnaireDeadline(any(LocalDateTime.class));
+        verify(directionsQuestionnaireDeadlineCalculator).calculateDirectionsQuestionnaireDeadline(eq(TODAY));
         verify(eventProducer).createClaimantResponseEvent(any(Claim.class), anyString());
-        verify(appInsights).trackEvent(eq(NON_LA_CASES), eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
+        verify(appInsights).trackEvent(eq(BOTH_PARTIES_OFFLINE_DQ), eq(REFERENCE_NUMBER),
+            eq(claim.getReferenceNumber()));
     }
 
     @Test
@@ -340,7 +357,8 @@ public class ClaimantResponseServiceTest {
         verify(caseRepository, never())
             .updateDirectionsQuestionnaireDeadline(any(Claim.class), any(LocalDate.class), anyString());
         verify(eventProducer).createClaimantResponseEvent(any(Claim.class), eq(AUTHORISATION));
-        verify(appInsights).trackEvent(eq(LA_PILOT_ELIGIBLE), eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
+        verify(appInsights).trackEvent(eq(BOTH_PARTIES_ONLINE_DQ), eq(REFERENCE_NUMBER),
+            eq(claim.getReferenceNumber()));
     }
 
     @Test
@@ -361,17 +379,15 @@ public class ClaimantResponseServiceTest {
         when(claimService.getClaimByExternalId(eq(EXTERNAL_ID), eq(AUTHORISATION))).thenReturn(claim);
         when(caseRepository.saveClaimantResponse(any(Claim.class), any(ResponseRejection.class), eq(AUTHORISATION)))
             .thenReturn(claim);
-        when(directionsQuestionnaireDeadlineCalculator
-            .calculateDirectionsQuestionnaireDeadline(any(LocalDateTime.class)))
+        when(directionsQuestionnaireDeadlineCalculator.calculateDirectionsQuestionnaireDeadline(eq(TODAY)))
             .thenReturn(dqDeadline);
 
         claimantResponseService.save(EXTERNAL_ID, claim.getSubmitterId(), claimantResponse, AUTHORISATION);
 
         verify(caseRepository).saveClaimantResponse(any(Claim.class), eq(claimantResponse), any());
-        verify(directionsQuestionnaireDeadlineCalculator)
-            .calculateDirectionsQuestionnaireDeadline(any(LocalDateTime.class));
+        verify(directionsQuestionnaireDeadlineCalculator).calculateDirectionsQuestionnaireDeadline(eq(TODAY));
         verify(eventProducer).createClaimantResponseEvent(any(Claim.class), eq(AUTHORISATION));
-        verify(appInsights).trackEvent(eq(NON_LA_CASES),
+        verify(appInsights).trackEvent(eq(BOTH_PARTIES_OFFLINE_DQ),
             eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
         verify(appInsights).trackEvent(eq(CLAIMANT_OPTED_OUT_FOR_MEDIATION_PILOT),
             eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
@@ -404,7 +420,7 @@ public class ClaimantResponseServiceTest {
         verify(caseRepository, never())
             .updateDirectionsQuestionnaireDeadline(any(Claim.class), any(LocalDate.class), anyString());
         verify(eventProducer).createClaimantResponseEvent(any(Claim.class), eq(AUTHORISATION));
-        verify(appInsights).trackEvent(eq(LA_PILOT_ELIGIBLE),
+        verify(appInsights).trackEvent(eq(BOTH_PARTIES_ONLINE_DQ),
             eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
         verify(appInsights).trackEvent(eq(CLAIMANT_OPTED_OUT_FOR_NON_MEDIATION_PILOT),
             eq(REFERENCE_NUMBER), eq(claim.getReferenceNumber()));
@@ -584,5 +600,71 @@ public class ClaimantResponseServiceTest {
 
         claimantResponseService.save(EXTERNAL_ID, claim.getSubmitterId(), claimantResponse, AUTHORISATION);
         verify(caseRepository, never()).saveCaseEvent(AUTHORISATION, claim, LIFT_STAY);
+    }
+
+    @Test
+    public void casesAssignedForJudgeDirectionsShouldRaiseJddoEligibleAppInsightsEvent() {
+        ClaimantResponse claimantResponse = SampleClaimantResponse.validDefaultRejection();
+
+        DirectionsQuestionnaire defendantHearingLocation =
+            SampleDirectionsQuestionnaire.builder()
+                .withHearingLocation(SampleHearingLocation.pilotHearingLocation)
+                .build();
+
+        FullDefenceResponse defendantResponse =
+            SampleResponse.FullDefence.builder()
+                .withDefenceType(DefenceType.ALREADY_PAID)
+                .withDirectionsQuestionnaire(defendantHearingLocation)
+                .build();
+
+        Claim claim = SampleClaim.builder()
+            .withResponse(defendantResponse)
+            .withRespondedAt(LocalDateTime.now().minusDays(34))
+            .withClaimantResponse(claimantResponse)
+            .withFeatures(ImmutableList.of(DQ_FLAG.getValue(), JUDGE_PILOT_FLAG.getValue()))
+            .build();
+
+        when(claimService.getClaimByExternalId(eq(EXTERNAL_ID), eq(AUTHORISATION))).thenReturn(claim);
+        when(caseRepository.saveClaimantResponse(any(Claim.class), any(), eq(AUTHORISATION)))
+            .thenReturn(claim);
+        when(directionsQuestionnaireService.prepareCaseEvent(any(), any()))
+            .thenReturn(Optional.of(ASSIGNING_FOR_JUDGE_DIRECTIONS));
+
+        claimantResponseService.save(EXTERNAL_ID, claim.getSubmitterId(), claimantResponse, AUTHORISATION);
+        verify(appInsights).trackEvent(eq(JDDO_PILOT_ELIGIBLE), eq(REFERENCE_NUMBER),
+            eq(claim.getReferenceNumber()));
+    }
+
+    @Test
+    public void casesAssignedForLADirectionsShouldRaiseLAPilotEligibleAppInsightsEvent() {
+        ClaimantResponse claimantResponse = SampleClaimantResponse.validDefaultRejection();
+
+        DirectionsQuestionnaire defendantHearingLocation =
+            SampleDirectionsQuestionnaire.builder()
+                .withHearingLocation(SampleHearingLocation.pilotHearingLocation)
+                .build();
+
+        FullDefenceResponse defendantResponse =
+            SampleResponse.FullDefence.builder()
+                .withDefenceType(DefenceType.ALREADY_PAID)
+                .withDirectionsQuestionnaire(defendantHearingLocation)
+                .build();
+
+        Claim claim = SampleClaim.builder()
+            .withResponse(defendantResponse)
+            .withRespondedAt(LocalDateTime.now().minusDays(34))
+            .withClaimantResponse(claimantResponse)
+            .withFeatures(ImmutableList.of(DQ_FLAG.getValue(), LA_PILOT_FLAG.getValue()))
+            .build();
+
+        when(claimService.getClaimByExternalId(eq(EXTERNAL_ID), eq(AUTHORISATION))).thenReturn(claim);
+        when(caseRepository.saveClaimantResponse(any(Claim.class), any(), eq(AUTHORISATION)))
+            .thenReturn(claim);
+        when(directionsQuestionnaireService.prepareCaseEvent(any(), any()))
+            .thenReturn(Optional.of(ASSIGNING_FOR_LEGAL_ADVISOR_DIRECTIONS));
+
+        claimantResponseService.save(EXTERNAL_ID, claim.getSubmitterId(), claimantResponse, AUTHORISATION);
+        verify(appInsights).trackEvent(eq(LA_PILOT_ELIGIBLE), eq(REFERENCE_NUMBER),
+            eq(claim.getReferenceNumber()));
     }
 }
